requirements.txt
pymupdf==1.26.4
easyocr==1.7.2
pillow==11.3.0
pylint==3.3.8
asyncpg==0.30.0
python-dotenv==1.1.1
pdf2image==1.17.0
tqdm==4.66.1